requirements.txt
# Cookbot dependencies
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
